requirements.txt
# LLM
openai>=1.0.0

# Browser
camoufox>=0.4.11
playwright>=1.40.0

# Database
asyncpg>=0.29.0

# HTML parsing
beautifulsoup4>=4.12.0
markdownify>=0.11.0

# HTTP
httpx>=0.25.0
curl_cffi>=0.7.0

# Search (Research Subagent)
duckduckgo-search>=6.0.0

# Config
python-dotenv>=1.0.0

# Dev / test
pytest>=7.0.0
pytest-asyncio>=0.21.0
